+++ src/gigachanie/eval/tasks/add-type-hints/repo/pt.py
@@ -1,0 +1,2 @@
+def scale(point, factor):
+    return (point[0] * factor, point[1] * factor)