+++ src/gigachanie/eval/tasks/fix-close-file/repo/io_util.py
@@ -1,0 +1,4 @@
+def load(path):
+    f = open(path)
+    data = f.read()
+    return data